requirements.txt
streamlit
langchain-openai
langchain-community
chromadb
pypdf
duckduckgo-search
pillow
requests
